Create styles.css
i think my issue is that the folder the name was wrong on the original file, i was trying to make a CSS folder and stick this stylles.css file into it. i think I accidently named the file itself CSS so I am reuploading it as I can't get bithub to let me change the name of the file.... or if there is a way I haven't found the right button to click an allow it yet.